Update users views, forms, tests after adding nickname
