+++ users/tests/test_registration.py
@@ -27,6 +27,7 @@ def test_registration_form_valid_data(
 ) -> None:
     data: dict = {
         "email": "testuser@example.com",
+        "nickname": "testuser",
         "password1": "Pass2712!",
         "password2": "Pass2712!",
     }
@@ -41,6 +42,7 @@ def test_registration_form_valid_data(
 def test_registration_form_missing_data(client: Client, user_model: type[User], register_url: str) -> None:
     data: dict = {
         "email": "",
+        "nickname": "",
         "password1": "",
         "password2": "",
     }
@@ -49,8 +51,9 @@ def test_registration_form_missing_data(client: Client, user_model: type[User], 
     assert "users/registration.html" in [t.name for t in response.templates]
     assert not user_model.objects.filter(email="").exists()
     form_errors = response.context["form"].errors
-    assert len(form_errors) == 3
+    assert len(form_errors) == 4
     assert "email" in form_errors
+    assert "nickname" in form_errors
     assert "password1" in form_errors
     assert "password2" in form_errors
 
@@ -65,6 +68,7 @@ def test_registration_form_user_already_exist(
 ) -> None:
     data: dict = {
         "email": user.email,
+        "nickname": user.nickname,
         "password1": generated_password,
         "password2": generated_password,
     }
@@ -73,6 +77,7 @@ def test_registration_form_user_already_exist(
     assert response.status_code == 200
     assert "users/registration.html" in [t.name for t in response.templates]
     form_errors = response.context["form"].errors
-    assert len(form_errors) == 1
+    assert len(form_errors) == 2
     assert "email" in form_errors
+    assert "nickname" in form_errors
     assert user_model.objects.count() == 1
