+++ users/tests/test_reset_password.py
@@ -20,13 +20,11 @@ def test_password_reset_view(
 @pytest.mark.django_db()
 def test_password_reset_not_send_email(
         client: Client,
-        generated_password: str,
+        user: type[User],
 ) -> None:
-    test_email = "test@example.com"
-    User.objects.create_user(email="testuser@example.com",
-                             password=generated_password)
+    wrong_email = "wrong@example.com"
     url = reverse("reset_password")
-    response = client.post(url, {"email": test_email})
+    response = client.post(url, {"email": wrong_email})
     assert response.status_code == 302
     assert len(mail.outbox) == 0
 
@@ -34,10 +32,8 @@ def test_password_reset_not_send_email(
 @pytest.mark.django_db()
 def test_password_reset_send_email(
         client: Client,
-        generated_password: str,
+        user: type[User],
 ) -> None:
-    test_email = "test@example.com"
-    user = User.objects.create_user(email=test_email, password=generated_password)
     url = reverse("reset_password")
     response = client.post(url, {"email": user.email})
     assert response.status_code == 302
@@ -47,10 +43,8 @@ def test_password_reset_send_email(
 @pytest.mark.django_db()
 def test_password_reset_twice_send_email(
         client: Client,
-        generated_password: str,
+        user: type[User],
 ) -> None:
-    test_email = "test@example.com"
-    user = User.objects.create_user(email=test_email, password=generated_password)
     url = reverse("reset_password")
     response = client.post(url, {"email": user.email})
     assert response.status_code == 302
@@ -65,11 +59,11 @@ def test_password_reset_twice_send_email(
 def test_password_first_wrong_mail_and_second_correct(
         client: Client,
         generated_password: str,
+        user: type[User],
 ) -> None:
-    test_email = "test1@example.com"
-    user = User.objects.create_user(email="test@example.com", password=generated_password)
+    wrong_email = "wrong@example.com"
     url = reverse("reset_password")
-    response = client.post(url, {"email": test_email})
+    response = client.post(url, {"email": wrong_email})
     assert response.status_code == 302
     assert len(mail.outbox) == 0
     url = reverse("reset_password")
