+++ users/tests/test_user_model.py
@@ -1,6 +1,5 @@
 import pytest
 from django.contrib.auth import get_user_model
-from django.contrib.auth.forms import AuthenticationForm
 from django.test import Client
 from django.urls import reverse_lazy
 
@@ -46,13 +45,17 @@ def test_create_superuser_missing_email(generated_password: str) -> None:
 @pytest.mark.django_db()
 def test_create_superuser_not_staff(generated_password: str) -> None:
     with pytest.raises(ValueError, match="Superuser must have is_staff=True."):
-        User.objects.create_superuser(email="admin@example.com", nickname="admin", password=generated_password, is_staff=False)
+        User.objects.create_superuser(
+            email="admin@example.com", nickname="admin", password=generated_password, is_staff=False,
+        )
 
 
 @pytest.mark.django_db()
 def test_create_superuser_not_superuser(generated_password: str) -> None:
     with pytest.raises(ValueError, match="Superuser must have is_superuser=True."):
-        User.objects.create_superuser(email="admin@example.com", nickname="admin", password=generated_password, is_superuser=False)
+        User.objects.create_superuser(
+            email="admin@example.com", nickname="admin", password=generated_password, is_superuser=False,
+        )
 
 
 def test_login_user_view_get(client: Client) -> None:
@@ -60,22 +63,3 @@ def test_login_user_view_get(client: Client) -> None:
     assert response.status_code == HTTP_SUCCESS
     assert b"Log in" in response.content
 
-
-@pytest.mark.django_db()
-def test_login_user_view_post_success_authenticated(client: Client, user: User) -> None:
-    data = {"email": "test@example.com", "password": "testpassword"}
-    response = client.post(reverse_lazy("login"), data=data)
-    assert response.status_code == HTTP_SUCCESS
-
-    user = User.objects.get(email="test@example.com")
-    assert user.is_authenticated
-
-
-def test_login_user_view_post_missing_credentials(client: Client) -> None:
-    data = {}
-    response = client.post(reverse_lazy("login"), data=data)
-    assert response.status_code == HTTP_SUCCESS
-    form = AuthenticationForm(data=data)
-    form.is_valid()
-    assert form.errors.get("username") == ["This field is required."]
-    assert form.errors.get("password") == ["This field is required."]
